answer get back - check chunks

+++ src/main/java/org/example/notebooklm/service/EmbeddingService.java
@@ -1,5 +1,5 @@
 package org.example.notebooklm.service;
 
 public interface EmbeddingService {
-    double[] generateEmbedding(String text);
+    float[] generateEmbedding(String text);
 }
